Added a couple more samples; moved models into a separate module
